Refactor: Split large components constants file
Split `src/lib/constants/components.ts` into individual component-specific files in `src/lib/constants/components/` to improve maintainability and navigation.
The original file now re-exports all constants from the new location, preserving backward compatibility.
Addressed unused imports and cleaned up comments.
