+++ src/lib/constants/components/accordion.ts
@@ -1,0 +1,25 @@
+/**
+ * Accordion-specific constants
+ */
+export const ACCORDION = {
+  SELECTORS: {
+    ACCORDION: '.c-accordion',
+    HEADER: '.c-accordion__header',
+    PANEL: '.c-accordion__panel',
+    BODY: '.c-accordion__body',
+  },
+  CLASSES: {
+    IS_OPEN: 'is-open',
+    IS_ANIMATING: 'is-animating',
+    IS_DISABLED: 'is-disabled',
+  },
+  ATTRIBUTES: {
+    ARIA_EXPANDED: 'aria-expanded',
+    ARIA_CONTROLS: 'aria-controls',
+    ARIA_HIDDEN: 'aria-hidden',
+    ROLE: 'role',
+  },
+  CSS_VARS: {
+    PANEL_HEIGHT: '--panel-height',
+  },
+};
