+++ src/lib/constants/components/avatar-group.ts
@@ -1,0 +1,14 @@
+/**
+ * Avatar Group-specific constants
+ */
+export const AVATAR_GROUP = {
+  SELECTORS: {
+    GROUP: '.c-avatar-group',
+    MORE: '.c-avatar-group__more',
+  },
+  CLASSES: {
+    BASE: 'c-avatar-group',
+    STACKED: 'c-avatar-group--stacked',
+    MORE: 'c-avatar-group__more',
+  },
+};
