+++ src/lib/constants/components/avatar.ts
@@ -1,0 +1,20 @@
+/**
+ * Avatar-specific constants
+ */
+export const AVATAR = {
+  SELECTORS: {
+    AVATAR: '.c-avatar',
+    IMAGE: '.c-avatar__image',
+    INITIALS: '.c-avatar__initials',
+    ICON: '.c-avatar__icon',
+  },
+  CLASSES: {
+    BASE: 'c-avatar',
+    XS: 'c-avatar--xs',
+    SM: 'c-avatar--sm',
+    MD: 'c-avatar--md',
+    LG: 'c-avatar--lg',
+    XL: 'c-avatar--xl',
+    CIRCLE: 'c-avatar--circle',
+  },
+};
