+++ src/lib/constants/components/badge.ts
@@ -1,0 +1,9 @@
+/**
+ * Badge-specific constants
+ */
+export const BADGE = {
+  BASE_CLASS: 'c-badge',
+  ICON_CLASS: 'c-badge__icon',
+  VARIANT_PREFIX: 'c-badge--',
+  SIZE_PREFIX: 'c-badge--',
+};
