+++ src/lib/constants/components/breadcrumb.ts
@@ -1,0 +1,19 @@
+/**
+ * Breadcrumb-specific constants
+ */
+export const BREADCRUMB = {
+  SELECTORS: {
+    BREADCRUMB: '.c-breadcrumb',
+    ITEM: '.c-breadcrumb__item',
+    LINK: '.c-breadcrumb__link',
+  },
+  CLASSES: {
+    BASE: 'c-breadcrumb',
+    ITEM: 'c-breadcrumb__item',
+    LINK: 'c-breadcrumb__link',
+    ACTIVE: 'is-active',
+  },
+  DEFAULTS: {
+    DIVIDER: '›',
+  },
+};
