+++ src/lib/constants/components/button.ts
@@ -1,0 +1,35 @@
+/**
+ * Button-specific constants
+ */
+export const THEME_NAMING = {
+  BUTTON_PREFIX: 'btn',
+  ICON_ELEMENT: 'icon',
+  LABEL_ELEMENT: 'label',
+  SPINNER_ELEMENT: 'spinner',
+};
+
+export const BUTTON = {
+  BASE_CLASS: 'c-btn',
+  ICON_CLASS: 'c-btn__icon',
+  LABEL_CLASS: 'c-btn__label',
+  SPINNER_CLASS: 'c-btn__spinner',
+  VARIANT_PREFIX: 'c-btn--',
+  CLASSES: {
+    BASE: 'c-btn',
+    LOADING: 'c-btn--loading',
+    FULL_WIDTH: 'c-btn--full-width',
+    BLOCK: 'c-btn--block',
+    ACTIVE: 'c-btn--active',
+    SELECTED: 'c-btn--selected',
+  },
+};
+
+/**
+ * ButtonGroup-specific constants
+ */
+export const BUTTON_GROUP = {
+  BASE_CLASS: 'c-btn-group',
+  CLASSES: {
+    BASE: 'c-btn-group',
+  },
+};
