+++ src/lib/constants/components/callout.ts
@@ -1,0 +1,19 @@
+/**
+ * Callout-specific constants
+ */
+export const CALLOUT = {
+  BASE_CLASS: 'c-callout',
+  CONTENT_CLASS: 'c-callout__content',
+  ICON_CLASS: 'c-callout__icon',
+  MESSAGE_CLASS: 'c-callout__message',
+  TITLE_CLASS: 'c-callout__title',
+  TEXT_CLASS: 'c-callout__text',
+  ACTIONS_CLASS: 'c-callout__actions',
+  CLOSE_BTN_CLASS: 'c-callout__close-btn',
+  VARIANT_PREFIX: 'c-callout--',
+  CLASSES: {
+    COMPACT: 'c-callout--compact',
+    TOAST: 'c-callout--toast',
+    HIDE: 'is-hide',
+  },
+};
