+++ src/lib/constants/components/chart.ts
@@ -1,0 +1,143 @@
+/**
+ * Chart-specific constants
+ */
+export const CHART = {
+  BASE_CLASS: 'c-chart',
+  ROOT_CLASS: 'c-chart',
+  HEADER_CLASS: 'c-chart__header',
+  HEADER_CONTENT_CLASS: 'c-chart__header-content',
+  TITLE_CLASS: 'c-chart__title',
+  SUBTITLE_CLASS: 'c-chart__subtitle',
+  TOOLBAR_CLASS: 'c-chart__toolbar',
+  ACTION_CLASS: 'c-chart__action',
+  EXPORT_GROUP_CLASS: 'c-chart__export-group',
+  EXPORT_DROPDOWN_CLASS: 'c-chart__export-dropdown',
+  EXPORT_OPTION_CLASS: 'c-chart__export-option',
+  SETTINGS_MENU_CLASS: 'c-chart__settings-menu',
+  TOOLBAR_GROUP_CLASS: 'c-chart__toolbar-group',
+  TOOLBAR_SEPARATOR_CLASS: 'c-chart__toolbar-separator',
+  TOOLBAR_LABEL_CLASS: 'c-chart__toolbar-label',
+  CONTENT_CLASS: 'c-chart__content',
+  CANVAS_CLASS: 'c-chart__canvas',
+  LEGEND_CLASS: 'c-chart__legend',
+  LEGEND_ITEM_CLASS: 'c-chart__legend-item',
+  LEGEND_LABEL_CLASS: 'c-chart__legend-label',
+  LEGEND_COLOR_CLASS: 'c-chart__legend-color',
+  TOOLTIP_CLASS: 'c-chart__tooltip',
+  TOOLTIP_TITLE_CLASS: 'c-chart__tooltip-title',
+  TOOLTIP_CONTENT_CLASS: 'c-chart__tooltip-content',
+  TOOLTIP_ITEM_CLASS: 'c-chart__tooltip-item',
+  TOOLTIP_LABEL_CLASS: 'c-chart__tooltip-label',
+  TOOLTIP_VALUE_CLASS: 'c-chart__tooltip-value',
+  AXIS_CLASS: 'c-chart__axis',
+  AXIS_LABEL_CLASS: 'c-chart__axis-label',
+  GRID_CLASS: 'c-chart__grid',
+  DATA_POINT_CLASS: 'c-chart__data-point',
+  CHART_SVG_CLASS: 'c-chart__svg',
+  LOADING_CLASS: 'c-chart__loading',
+  LOADING_SPINNER_CLASS: 'c-chart__loading-spinner',
+  LOADING_TEXT_CLASS: 'c-chart__loading-text',
+  ERROR_CLASS: 'c-chart__error',
+  ERROR_ICON_CLASS: 'c-chart__error-icon',
+  ERROR_CONTENT_CLASS: 'c-chart__error-content',
+  ERROR_MESSAGE_CLASS: 'c-chart__error-message',
+  ERROR_DETAILS_CLASS: 'c-chart__error-details',
+  EMPTY_CLASS: 'c-chart__empty',
+  EMPTY_ICON_CLASS: 'c-chart__empty-icon',
+  EMPTY_MESSAGE_CLASS: 'c-chart__empty-message',
+  EMPTY_DETAILS_CLASS: 'c-chart__empty-details',
+  DEFAULT_WIDTH: 800,
+  DEFAULT_HEIGHT: 400,
+  PIE_RADIUS_RATIO: 0.8,
+  DEFAULT_COLORS: [
+    'var(--atomix-primary)',
+    'var(--atomix-secondary)',
+    'var(--atomix-success)',
+    'var(--atomix-info)',
+    'var(--atomix-warning)',
+    'var(--atomix-error)',
+    'var(--atomix-primary-5)',
+    'var(--atomix-primary-7)',
+    'var(--atomix-primary-3)',
+    'var(--atomix-gray-6)',
+    'var(--atomix-gray-8)',
+    'var(--atomix-gray-4)',
+  ],
+  TYPE_PREFIX: 'c-chart--',
+  SIZE_PREFIX: 'c-chart--',
+  VARIANT_PREFIX: 'c-chart--',
+  LOADING_STATE_CLASS: 'c-chart--loading',
+  ERROR_STATE_CLASS: 'c-chart--error',
+  CLASSES: {
+    // Chart Types
+    LINE: 'c-chart--line',
+    AREA: 'c-chart--area',
+    BAR: 'c-chart--bar',
+    HORIZONTAL_BAR: 'c-chart--horizontal-bar',
+    PIE: 'c-chart--pie',
+    DONUT: 'c-chart--donut',
+    DOUGHNUT: 'c-chart--doughnut',
+    SCATTER: 'c-chart--scatter',
+    RADAR: 'c-chart--radar',
+    BUBBLE: 'c-chart--bubble',
+    CANDLESTICK: 'c-chart--candlestick',
+    GAUGE: 'c-chart--gauge',
+    FUNNEL: 'c-chart--funnel',
+    WATERFALL: 'c-chart--waterfall',
+    HEATMAP: 'c-chart--heatmap',
+    TREEMAP: 'c-chart--treemap',
+
+    // Sizes
+    SM: 'c-chart--sm',
+    MD: 'c-chart--md',
+    LG: 'c-chart--lg',
+    XL: 'c-chart--xl',
+    FULL: 'c-chart--full',
+
+    // Variants
+    PRIMARY: 'c-chart--primary',
+    SECONDARY: 'c-chart--secondary',
+    SUCCESS: 'c-chart--success',
+    ERROR: 'c-chart--error',
+    WARNING: 'c-chart--warning',
+    INFO: 'c-chart--info',
+
+    // States
+    LOADING: 'c-chart--loading',
+    INTERACTIVE: 'c-chart--interactive',
+    DISABLED: 'c-chart--disabled',
+    FULLSCREEN: 'c-chart--fullscreen',
+    MINIMIZED: 'c-chart--minimized',
+
+    // Legacy Support
+    ADVANCED: 'c-chart--advanced',
+
+    // Toolbar sizes
+    TOOLBAR_SM: 'c-chart__toolbar--sm',
+    TOOLBAR_MD: 'c-chart__toolbar--md',
+    TOOLBAR_LG: 'c-chart__toolbar--lg',
+
+    // Toolbar positions
+    TOOLBAR_TOP: 'c-chart__toolbar--top',
+    TOOLBAR_BOTTOM: 'c-chart__toolbar--bottom',
+    TOOLBAR_LEFT: 'c-chart__toolbar--left',
+    TOOLBAR_RIGHT: 'c-chart__toolbar--right',
+
+    // Action states
+    ACTION_ACTIVE: 'c-chart__action--active',
+    ACTION_DISABLED: 'c-chart__action--disabled',
+    ACTION_PRIMARY: 'c-chart__action--primary',
+    ACTION_SECONDARY: 'c-chart__action--secondary',
+    ACTION_SUCCESS: 'c-chart__action--success',
+    ACTION_INFO: 'c-chart__action--info',
+    ACTION_WARNING: 'c-chart__action--warning',
+    ACTION_ERROR: 'c-chart__action--error',
+  },
+
+  // Toolbar configuration
+  TOOLBAR: {
+    SIZES: ['sm', 'md', 'lg'] as const,
+    POSITIONS: ['top', 'bottom', 'left', 'right'] as const,
+    ACTION_VARIANTS: ['primary', 'secondary', 'success', 'info', 'warning', 'error'] as const,
+  },
+};
