+++ src/lib/constants/components/countdown.ts
@@ -1,0 +1,20 @@
+/**
+ * Countdown-specific constants
+ */
+export const COUNTDOWN = {
+  SELECTORS: {
+    COUNTDOWN: '.c-countdown',
+    TIME: '.c-countdown__time',
+    TIME_COUNT: '.c-countdown__time-count',
+    TIME_LABEL: '.c-countdown__time-label',
+    SEPARATOR: '.c-countdown__separator',
+  },
+  CLASSES: {
+    BASE: 'c-countdown',
+    FOCUSED: 'c-countdown--focused',
+  },
+  DEFAULTS: {
+    SEPARATOR: ':',
+    SHOW: ['days', 'hours', 'minutes', 'seconds'],
+  },
+};
