+++ src/lib/constants/components/data-table.ts
@@ -1,0 +1,51 @@
+/**
+ * DataTable-specific constants
+ */
+export const DATA_TABLE_CLASSES = {
+  base: 'c-data-table',
+  container: 'c-data-table-container',
+  tableWrapper: 'c-data-table-wrapper',
+  header: 'c-data-table__header',
+  headerCell: 'c-data-table__header-cell',
+  headerContent: 'c-data-table__header-content',
+  headerActions: 'c-data-table__header-actions',
+  sortable: 'c-data-table__header-cell--sortable',
+  sortIcon: 'c-data-table__sort-icon',
+  row: 'c-data-table__row',
+  rowSelected: 'c-data-table__row--selected',
+  cell: 'c-data-table__cell',
+  selectionCell: 'c-data-table__cell--selection',
+  loadingCell: 'c-data-table__loading-cell',
+  loadingIndicator: 'c-data-table__loading-indicator',
+  emptyCell: 'c-data-table__empty-cell',
+  toolbar: 'c-data-table-toolbar',
+  toolbarLeft: 'c-data-table-toolbar__left',
+  toolbarRight: 'c-data-table-toolbar__right',
+  search: 'c-data-table-search',
+  searchInput: 'c-data-table-search__input',
+  columnFilter: 'c-data-table__column-filter',
+  pagination: 'c-data-table__pagination-container',
+  striped: 'c-data-table--striped',
+  bordered: 'c-data-table--bordered',
+  dense: 'c-data-table--dense',
+  loading: 'c-data-table--loading',
+  stickyHeader: 'c-data-table--sticky-header',
+  dragging: 'c-data-table__header-cell--dragging',
+  dragOver: 'c-data-table__header-cell--drag-over',
+  resizeHandle: 'c-data-table__resize-handle',
+  open: 'is-open',
+};
+
+/**
+ * DataTable-specific selectors
+ */
+export const DATA_TABLE_SELECTORS = {
+  TABLE: '.c-data-table',
+  HEADER: '.c-data-table__header',
+  HEADER_CELL: '.c-data-table__header-cell',
+  ROW: '.c-data-table__row',
+  CELL: '.c-data-table__cell',
+  PAGINATION: '.c-data-table__pagination',
+  PAGINATION_BUTTON: '.c-data-table__pagination-button',
+  SEARCH_INPUT: '.c-data-table__search-input',
+};
