+++ src/lib/constants/components/footer.ts
@@ -1,0 +1,87 @@
+/**
+ * Footer-specific constants
+ */
+export const FOOTER = {
+  SELECTORS: {
+    FOOTER: '.c-footer',
+    CONTAINER: '.c-footer__container',
+    SECTIONS: '.c-footer__sections',
+    BRAND: '.c-footer__brand',
+    BRAND_LOGO: '.c-footer__brand-logo',
+    BRAND_NAME: '.c-footer__brand-name',
+    BRAND_DESCRIPTION: '.c-footer__brand-description',
+    SECTION: '.c-footer__section',
+    SECTION_HEADER: '.c-footer__section-header',
+    SECTION_TITLE: '.c-footer__section-title',
+    SECTION_CONTENT: '.c-footer__section-content',
+    SECTION_TOGGLE: '.c-footer__section-toggle',
+    LINK: '.c-footer__link',
+    SOCIAL: '.c-footer__social',
+    SOCIAL_LINK: '.c-footer__social-link',
+    NEWSLETTER: '.c-footer__newsletter',
+    NEWSLETTER_FORM: '.c-footer__newsletter-form',
+    NEWSLETTER_INPUT: '.c-footer__newsletter-input',
+    NEWSLETTER_BUTTON: '.c-footer__newsletter-button',
+    BOTTOM: '.c-footer__bottom',
+    COPYRIGHT: '.c-footer__copyright',
+    BACK_TO_TOP: '.c-footer__back-to-top',
+    DIVIDER: '.c-footer__divider',
+  },
+  CLASSES: {
+    BASE: 'c-footer',
+    CONTAINER: 'c-footer__container',
+    SECTIONS: 'c-footer__sections',
+    BRAND: 'c-footer__brand',
+    BRAND_LOGO: 'c-footer__brand-logo',
+    BRAND_NAME: 'c-footer__brand-name',
+    BRAND_DESCRIPTION: 'c-footer__brand-description',
+    SECTION: 'c-footer__section',
+    SECTION_HEADER: 'c-footer__section-header',
+    SECTION_TITLE: 'c-footer__section-title',
+    SECTION_CONTENT: 'c-footer__section-content',
+    SECTION_TOGGLE: 'c-footer__section-toggle',
+    SECTION_COLLAPSIBLE: 'c-footer__section--collapsible',
+    SECTION_COLLAPSED: 'c-footer__section--collapsed',
+    LINK: 'c-footer__link',
+    LINK_ACTIVE: 'c-footer__link--active',
+    LINK_DISABLED: 'c-footer__link--disabled',
+    SOCIAL: 'c-footer__social',
+    SOCIAL_LINK: 'c-footer__social-link',
+    NEWSLETTER: 'c-footer__newsletter',
+    NEWSLETTER_FORM: 'c-footer__newsletter-form',
+    NEWSLETTER_INPUT: 'c-footer__newsletter-input',
+    NEWSLETTER_BUTTON: 'c-footer__newsletter-button',
+    BOTTOM: 'c-footer__bottom',
+    COPYRIGHT: 'c-footer__copyright',
+    BACK_TO_TOP: 'c-footer__back-to-top',
+    DIVIDER: 'c-footer__divider',
+    // Layout variants
+    COLUMNS: 'c-footer--columns',
+    CENTERED: 'c-footer--centered',
+    MINIMAL: 'c-footer--minimal',
+    STACKED: 'c-footer--stacked',
+    FLEXIBLE: 'c-footer--flexible',
+    SIDEBAR: 'c-footer--sidebar',
+    WIDE: 'c-footer--wide',
+    // Size variants
+    SM: 'c-footer--sm',
+    MD: 'c-footer--md',
+    LG: 'c-footer--lg',
+    // State modifiers
+    STICKY: 'c-footer--sticky',
+  },
+  DEFAULTS: {
+    LAYOUT: 'columns',
+    VARIANT: 'primary',
+    SIZE: 'md',
+    SHOW_NEWSLETTER: false,
+    SHOW_BACK_TO_TOP: false,
+    SHOW_DIVIDER: true,
+    STICKY: false,
+    NEWSLETTER_TITLE: 'Stay Updated',
+    NEWSLETTER_DESCRIPTION: 'Subscribe to our newsletter for the latest updates.',
+    NEWSLETTER_PLACEHOLDER: 'Enter your email',
+    NEWSLETTER_BUTTON_TEXT: 'Subscribe',
+    BACK_TO_TOP_TEXT: 'Back to Top',
+  },
+};
