+++ src/lib/constants/components/form-group.ts
@@ -1,0 +1,20 @@
+/**
+ * Form Group-specific constants
+ */
+export const FORM_GROUP = {
+  SELECTORS: {
+    GROUP: '.c-form-group',
+    LABEL: '.c-form-group__label',
+    FIELD: '.c-form-group__field',
+    HELPER: '.c-form-group__helper',
+    REQUIRED: '.c-form-group__required',
+  },
+  CLASSES: {
+    BASE: 'c-form-group',
+    SMALL: 'c-form-group--sm',
+    LARGE: 'c-form-group--lg',
+    INVALID: 'c-form-group--invalid',
+    VALID: 'c-form-group--valid',
+    DISABLED: 'c-form-group--disabled',
+  },
+};
