+++ src/lib/constants/components/form.ts
@@ -1,0 +1,17 @@
+/**
+ * Form-specific constants
+ */
+export const FORM = {
+  SELECTORS: {
+    FORM: '.c-form',
+    GROUP: '.c-form-group',
+    LABEL: '.c-form-group__label',
+    HELPER: '.c-form-group__helper',
+    FIELD: '.c-form-group__field',
+    REQUIRED: '.c-form-group__required',
+  },
+  CLASSES: {
+    BASE: 'c-form',
+    DISABLED: 'c-form--disabled',
+  },
+};
