+++ src/lib/constants/components/hero.ts
@@ -1,0 +1,32 @@
+/**
+ * Hero-specific constants
+ */
+export const HERO = {
+  SELECTORS: {
+    HERO: '.c-hero',
+    CONTAINER: '.c-hero__container',
+    GRID: '.c-hero__grid',
+    CONTENT: '.c-hero__content',
+    SUBTITLE: '.c-hero__subtitle',
+    TITLE: '.c-hero__title',
+    TEXT: '.c-hero__text',
+    ACTIONS: '.c-hero__actions',
+    IMAGE: '.c-hero__image',
+    BG: '.c-hero__bg',
+    BG_IMAGE: '.c-hero__bg-image',
+    OVERLAY: '.c-hero__overlay',
+    IMAGE_WRAPPER: '.c-hero__image-wrapper',
+    SLIDER: '.c-hero__slider',
+    SLIDER_ITEM: '.c-hero__slider-item',
+  },
+  CLASSES: {
+    CENTER: 'c-hero--center',
+    RIGHT: 'c-hero--right',
+    LEFT: 'c-hero--left',
+    FULL_VH: 'c-hero--full-vh',
+    SLIDER_FADE: 'c-hero__slider--fade',
+    SLIDER_SLIDE: 'c-hero__slider--slide',
+    SLIDER_CUSTOM: 'c-hero__slider--custom',
+    SLIDER_ITEM_ACTIVE: 'c-hero__slider-item--active',
+  },
+};
