+++ src/lib/constants/components/input.ts
@@ -1,0 +1,32 @@
+/**
+ * Input-specific constants
+ */
+export const INPUT = {
+  SELECTORS: {
+    INPUT: '.c-input',
+  },
+  CLASSES: {
+    BASE: 'c-input',
+    SMALL: 'c-input--sm',
+    LARGE: 'c-input--lg',
+    INVALID: 'is-invalid',
+    VALID: 'is-valid',
+    DISABLED: 'is-disabled',
+    FULL_WIDTH: 'c-input--full-width',
+    PREFIX_ICON: 'c-input--prefix-icon',
+    SUFFIX_ICON: 'c-input--suffix-icon',
+    CLEARABLE: 'c-input--clearable',
+    WITH_COUNTER: 'c-input--with-counter',
+    PASSWORD_TOGGLE: 'c-input--password-toggle',
+  },
+  ELEMENTS: {
+    WRAPPER: 'c-input-wrapper',
+    PREFIX: 'c-input__prefix',
+    SUFFIX: 'c-input__suffix',
+    CLEAR_BUTTON: 'c-input__clear',
+    PASSWORD_TOGGLE: 'c-input__password-toggle',
+    COUNTER: 'c-input__counter',
+    ERROR_MESSAGE: 'c-input__error',
+    HELPER_TEXT: 'c-input__helper',
+  },
+};
