+++ src/lib/constants/components/list.ts
@@ -1,0 +1,23 @@
+/**
+ * List-specific constants
+ */
+export const LIST = {
+  BASE_CLASS: 'c-list',
+  ITEM_CLASS: 'c-list__item',
+  VARIANT_PREFIX: 'c-list--',
+  SIZE_PREFIX: 'c-list--',
+  CLASSES: {
+    ORDERED: 'c-list--ordered',
+    INLINE: 'c-list--inline',
+  },
+};
+
+/**
+ * List Group-specific constants
+ */
+export const LIST_GROUP = {
+  BASE_CLASS: 'c-list-group',
+  ITEM_CLASS: 'c-list-group__item',
+  VARIANT_PREFIX: 'c-list-group--',
+  SIZE_PREFIX: 'c-list-group--',
+};
