+++ src/lib/constants/components/modal.ts
@@ -1,0 +1,23 @@
+/**
+ * Modal component constants
+ */
+export const MODAL = {
+  SELECTORS: {
+    MODAL: '.c-modal',
+    OPEN_BUTTON: '.js-modal-open',
+    CLOSE_BUTTONS: '.js-modal-close',
+    DIALOG: '.c-modal__dialog',
+    BACKDROP: '.c-modal__backdrop',
+  },
+  CLASSES: {
+    IS_OPEN: 'is-open',
+  },
+  DEFAULT_OPTIONS: {
+    openELm: '.js-modal-open',
+    closeELms: '.js-modal-close',
+    modalDialogELm: '.c-modal__dialog',
+    backdropELm: '.c-modal__backdrop',
+    backdrop: true,
+    keyboard: true,
+  },
+};
