+++ src/lib/constants/components/nav.ts
@@ -1,0 +1,20 @@
+/**
+ * Nav-specific constants
+ */
+export const NAV = {
+  SELECTORS: {
+    NAV: '.c-nav',
+    ITEM: '.c-nav__item',
+    LINK: '.c-nav__link',
+    DROPDOWN: '.c-nav__item--dropdown',
+    DROPDOWN_MENU: '.c-nav__dropdown-menu',
+    MEGA_MENU: '.c-nav__mega-menu',
+    ICON: '.c-nav__icon',
+  },
+  CLASSES: {
+    END: 'c-nav--end',
+    CENTER: 'c-nav--center',
+    ACTIVE: 'is-active',
+    DISABLED: 'is-disabled',
+  },
+};
