+++ src/lib/constants/components/pagination.ts
@@ -1,0 +1,13 @@
+import { Size } from '../../types/components';
+
+/**
+ * Pagination-specific constants
+ */
+export const PAGINATION_DEFAULTS = {
+  currentPage: 1,
+  totalPages: 1,
+  siblingCount: 1,
+  showFirstLastButtons: true,
+  showPrevNextButtons: true,
+  size: 'md' as Size,
+};
