+++ src/lib/constants/components/photo-viewer.ts
@@ -1,0 +1,12 @@
+/**
+ * PhotoViewer component constants
+ */
+export const PHOTOVIEWER = {
+  SELECTOR: '.c-photo-viewer',
+  CLASS: 'c-photo-viewer',
+  DEFAULTS: {
+    startIndex: 0,
+    zoomLevel: 1,
+    fullscreen: false,
+  },
+};
