+++ src/lib/constants/components/progress.ts
@@ -1,0 +1,28 @@
+/**
+ * Progress-specific constants
+ */
+export const PROGRESS = {
+  SELECTORS: {
+    PROGRESS: '.c-progress',
+    BAR: '.c-progress__bar',
+  },
+  CLASSES: {
+    BASE: 'c-progress',
+    BAR: 'c-progress__bar',
+    SM: 'c-progress--sm',
+    MD: 'c-progress--md',
+    LG: 'c-progress--lg',
+  },
+  ATTRIBUTES: {
+    ARIA_VALUEMIN: 'aria-valuemin',
+    ARIA_VALUEMAX: 'aria-valuemax',
+    ARIA_VALUENOW: 'aria-valuenow',
+    ARIA_LABEL: 'aria-label',
+  },
+  CSS_VARS: {
+    PERCENTAGE: '--atomix-progress-percentage',
+  },
+  DEFAULTS: {
+    ARIA_LABEL: 'Progress bar',
+  },
+};
