+++ src/lib/constants/components/radio.ts
@@ -1,0 +1,16 @@
+/**
+ * Radio-specific constants
+ */
+export const RADIO = {
+  SELECTORS: {
+    RADIO: '.c-radio',
+    INPUT: '.c-radio__input',
+    LABEL: '.c-radio__label',
+  },
+  CLASSES: {
+    BASE: 'c-radio',
+    INVALID: 'is-error',
+    VALID: 'is-valid',
+    DISABLED: 'is-disabled',
+  },
+};
