+++ src/lib/constants/components/rating.ts
@@ -1,0 +1,21 @@
+/**
+ * Rating-specific constants
+ */
+export const RATING = {
+  SELECTORS: {
+    RATING: '.c-rating',
+    STAR: '.c-rating__star',
+    STAR_FULL: '.c-rating__star-full',
+    STAR_HALF: '.c-rating__star-half',
+  },
+  CLASSES: {
+    FULL: 'c-rating__star--full',
+    HALF: 'c-rating__star--half',
+    SMALL: 'c-rating--sm',
+    LARGE: 'c-rating--lg',
+  },
+  ATTRIBUTES: {
+    READONLY: 'data-readonly',
+    VALUE: 'data-value',
+  },
+};
