+++ src/lib/constants/components/river.ts
@@ -1,0 +1,31 @@
+/**
+ * River-specific constants
+ */
+export const RIVER = {
+  SELECTORS: {
+    RIVER: '.c-river',
+    CONTAINER: '.c-river__container',
+    ROW: '.c-river__row',
+    CONTENT: '.c-river__content',
+    CONTENT_COL: '.c-river__content-col',
+    CONTENT_COL_TITLE: '.c-river__content-col--title',
+    CONTENT_COL_TEXT: '.c-river__content-col--text',
+    TITLE: '.c-river__title',
+    TEXT: '.c-river__text',
+    ACTIONS: '.c-river__actions',
+    VISUAL: '.c-river__visual',
+    IMAGE_WRAPPER: '.c-river__image-wrapper',
+    IMAGE: '.c-river__image',
+    BG: '.c-river__bg',
+    BG_IMAGE: '.c-river__bg-image',
+    OVERLAY: '.c-river__overlay',
+  },
+  CLASSES: {
+    CENTER: 'c-river--center',
+    BREAKOUT: 'c-river--breakout',
+    REVERSE: 'c-river--reverse',
+  },
+  ATTRIBUTES: {
+    CONTENT_WIDTH: '--river-content-width',
+  },
+};
