+++ src/lib/constants/components/section-intro.ts
@@ -1,0 +1,17 @@
+/**
+ * SectionIntro-specific constants
+ */
+export const SECTION_INTRO = {
+  SELECTORS: {
+    SECTION_INTRO: '.c-sectionintro',
+    LABEL: '.c-sectionintro__label',
+    TITLE: '.c-sectionintro__title',
+    TEXT: '.c-sectionintro__text',
+    ACTIONS: '.c-sectionintro__actions',
+  },
+  CLASSES: {
+    CENTER: 'c-sectionintro--center',
+    LARGE: 'c-sectionintro--lg',
+    SMALL: 'c-sectionintro--sm',
+  },
+};
