+++ src/lib/constants/components/select.ts
@@ -1,0 +1,32 @@
+/**
+ * Select-specific constants
+ */
+export const SELECT = {
+  SELECTORS: {
+    SELECT: '.c-select',
+    SELECTED: '.c-select__selected',
+    SELECT_BODY: '.c-select__body',
+    SELECT_PANEL: '.c-select__panel',
+    SELECT_ITEMS: '.c-select__items',
+    SELECT_ITEM: '.c-select__item',
+    ITEM_LABEL: '.c-select__item-label',
+    ITEM_INPUT: '.c-select__item-input',
+    OPTION: 'option',
+  },
+  CLASSES: {
+    BASE: 'c-select',
+    SELECTED: 'c-select__selected',
+    SELECT_BODY: 'c-select__body',
+    SELECT_PANEL: 'c-select__panel',
+    SELECT_ITEMS: 'c-select__items',
+    SELECT_ITEM: 'c-select__item',
+    TOGGLE_ICON: 'c-select__toggle-icon',
+    ICON_CARET: 'icon-atomix-caret-down',
+    SMALL: 'c-select--sm',
+    LARGE: 'c-select--lg',
+    INVALID: 'is-invalid',
+    VALID: 'is-valid',
+    DISABLED: 'is-disabled',
+    IS_OPEN: 'is-open',
+  },
+};
