+++ src/lib/constants/components/side-menu.ts
@@ -1,0 +1,45 @@
+/**
+ * SideMenu-specific constants
+ */
+export const SIDE_MENU = {
+  SELECTORS: {
+    SIDE_MENU: '.c-side-menu',
+    WRAPPER: '.c-side-menu__wrapper',
+    INNER: '.c-side-menu__inner',
+    TITLE: '.c-side-menu__title',
+    TOGGLER: '.c-side-menu__toggler',
+    TOGGLER_ICON: '.c-side-menu__toggler-icon',
+    LIST: '.c-side-menu__list',
+    ITEM: '.c-side-menu__item',
+    LINK: '.c-side-menu__link',
+    LINK_ICON: '.c-side-menu__link-icon',
+    LINK_TEXT: '.c-side-menu__link-text',
+  },
+  CLASSES: {
+    BASE: 'c-side-menu',
+    WRAPPER: 'c-side-menu__wrapper',
+    INNER: 'c-side-menu__inner',
+    TITLE: 'c-side-menu__title',
+    TOGGLER: 'c-side-menu__toggler',
+    TOGGLER_ICON: 'c-side-menu__toggler-icon',
+    LIST: 'c-side-menu__list',
+    ITEM: 'c-side-menu__item',
+    LINK: 'c-side-menu__link',
+    LINK_ICON: 'c-side-menu__link-icon',
+    LINK_TEXT: 'c-side-menu__link-text',
+    IS_OPEN: 'is-open',
+    ACTIVE: 'is-active',
+    DISABLED: 'is-disabled',
+  },
+  ATTRIBUTES: {
+    SIDE_MENU: 'data-side-menu',
+    COLLAPSIBLE: 'data-collapsible',
+    OPEN: 'data-open',
+    TITLE: 'data-title',
+  },
+  DEFAULTS: {
+    COLLAPSIBLE: true,
+    OPEN: false,
+    TOGGLE_ICON: '▶',
+  },
+};
