+++ src/lib/constants/components/slider.ts
@@ -1,0 +1,92 @@
+/**
+ * Slider component constants
+ */
+export const sliderConstants = {
+  classes: {
+    container: 'c-slider',
+    containerModifierClass: 'c-slider--',
+    wrapper: 'c-slider__wrapper',
+    slide: 'c-slider__slide',
+    slideActive: 'c-slider__slide--active',
+    slidePrev: 'c-slider__slide--prev',
+    slideNext: 'c-slider__slide--next',
+    slideDuplicate: 'c-slider__slide--duplicate',
+    slideVisible: 'c-slider__slide--visible',
+    pagination: 'c-slider__pagination',
+    paginationBullet: 'c-slider__pagination-bullet',
+    paginationBulletActive: 'c-slider__pagination-bullet--active',
+    navigation: 'c-slider__navigation',
+    navigationPrev: 'c-slider__navigation-prev',
+    navigationNext: 'c-slider__navigation-next',
+    navigationDisabled: 'c-slider__navigation--disabled',
+    scrollbar: 'c-slider__scrollbar',
+    scrollbarDrag: 'c-slider__scrollbar-drag',
+    thumbs: 'c-slider__thumbs',
+    thumbsWrapper: 'c-slider__thumbs-wrapper',
+    thumbsSlide: 'c-slider__thumbs-slide',
+    thumbsSlideActive: 'c-slider__thumbs-slide--active',
+    zoomContainer: 'c-slider__zoom-container',
+    lazyPreloader: 'c-slider__lazy-preloader',
+  },
+  defaults: {
+    slidesToShow: 1,
+    slidesToScroll: 1,
+    spaceBetween: 0,
+    centeredSlides: false,
+    loop: false,
+    initialSlide: 0,
+    direction: 'horizontal' as const,
+    speed: 300,
+    easing: 'ease-out',
+    allowTouchMove: true,
+    threshold: 5,
+    mousewheel: false,
+    keyboard: false,
+    grabCursor: false,
+    freeMode: false,
+    watchSlidesProgress: false,
+    watchOverflow: true,
+    resistanceRatio: 0.85,
+    preventClicks: true,
+    preventClicksPropagation: true,
+    simulateTouch: true,
+    touchRatio: 1,
+    touchAngle: 45,
+    shortSwipes: true,
+    longSwipes: true,
+    longSwipesRatio: 0.5,
+    longSwipesMs: 300,
+    followFinger: true,
+    touchMoveStopPropagation: true,
+    touchStartPreventDefault: true,
+    touchReleaseOnEdges: false,
+    resistance: true,
+    passiveListeners: true,
+  },
+  breakpoints: {
+    sm: 576,
+    md: 768,
+    lg: 992,
+    xl: 1200,
+    xxl: 1400,
+  },
+  events: {
+    init: 'slider:init',
+    destroy: 'slider:destroy',
+    slideChange: 'slider:slideChange',
+    slideChangeTransitionStart: 'slider:slideChangeTransitionStart',
+    slideChangeTransitionEnd: 'slider:slideChangeTransitionEnd',
+    touchStart: 'slider:touchStart',
+    touchMove: 'slider:touchMove',
+    touchEnd: 'slider:touchEnd',
+    reachBeginning: 'slider:reachBeginning',
+    reachEnd: 'slider:reachEnd',
+    progress: 'slider:progress',
+    autoplayStart: 'slider:autoplayStart',
+    autoplayStop: 'slider:autoplayStop',
+    beforeResize: 'slider:beforeResize',
+    resize: 'slider:resize',
+  },
+};
+
+export const SLIDER = sliderConstants;
