+++ src/lib/constants/components/spinner.ts
@@ -1,0 +1,21 @@
+/**
+ * Spinner-specific constants
+ */
+export const SPINNER = {
+  SELECTORS: {
+    SPINNER: '.c-spinner',
+  },
+  CLASSES: {
+    PRIMARY: 'c-spinner--primary',
+    SECONDARY: 'c-spinner--secondary',
+    SUCCESS: 'c-spinner--success',
+    INFO: 'c-spinner--info',
+    WARNING: 'c-spinner--warning',
+    DANGER: 'c-spinner--danger',
+    LIGHT: 'c-spinner--light',
+    DARK: 'c-spinner--dark',
+    SMALL: 'c-spinner--sm',
+    LARGE: 'c-spinner--lg',
+  },
+  VISUALLY_HIDDEN: 'u-visually-hidden',
+};
