+++ src/lib/constants/components/steps.ts
@@ -1,0 +1,18 @@
+/**
+ * Steps-specific constants
+ */
+export const STEPS = {
+  SELECTORS: {
+    STEPS: '.c-steps',
+    ITEM: '.c-steps__item',
+    LINE: '.c-steps__line',
+    CONTENT: '.c-steps__content',
+    NUMBER: '.c-steps__number',
+    TEXT: '.c-steps__text',
+  },
+  CLASSES: {
+    ACTIVE: 'is-active',
+    VERTICAL: 'c-steps--vertical',
+    COMPLETED: 'is-completed',
+  },
+};
