+++ src/lib/constants/components/tab.ts
@@ -1,0 +1,18 @@
+/**
+ * Tab-specific constants
+ */
+export const TAB = {
+  SELECTORS: {
+    TAB: '.js-atomix-tab',
+    NAV_ITEMS: '.c-tabs__nav-item',
+    NAV_BTN: '.c-tabs__nav-btn',
+    PANELS: '.c-tabs__panel',
+    PANEL_BODIES: '.c-tabs__panel-body',
+  },
+  CLASSES: {
+    ACTIVE: 'is-active',
+  },
+  DEFAULTS: {
+    ACTIVE_INDEX: 0,
+  },
+};
