+++ src/lib/constants/components/testimonial.ts
@@ -1,0 +1,18 @@
+/**
+ * Testimonial-specific constants
+ */
+export const TESTIMONIAL = {
+  SELECTORS: {
+    TESTIMONIAL: '.c-testimonial',
+    QUOTE: '.c-testimonial__quote',
+    AUTHOR: '.c-testimonial__author',
+    AUTHOR_AVATAR: '.c-testimonial__author-avatar',
+    AUTHOR_INFO: '.c-testimonial__info',
+    AUTHOR_NAME: '.c-testimonial__author-name',
+    AUTHOR_ROLE: '.c-testimonial__author-role',
+  },
+  CLASSES: {
+    SMALL: 'c-testimonial--sm',
+    LARGE: 'c-testimonial--lg',
+  },
+};
