+++ src/lib/constants/components/textarea.ts
@@ -1,0 +1,16 @@
+/**
+ * Textarea-specific constants
+ */
+export const TEXTAREA = {
+  SELECTORS: {
+    TEXTAREA: '.c-textarea',
+  },
+  CLASSES: {
+    BASE: 'c-input c-input--textarea',
+    SMALL: 'c-input--sm',
+    LARGE: 'c-input--lg',
+    INVALID: 'is-invalid',
+    VALID: 'is-valid',
+    DISABLED: 'is-disabled',
+  },
+};
