+++ src/lib/constants/components/todo.ts
@@ -1,0 +1,26 @@
+/**
+ * Todo-specific constants
+ */
+export const TODO = {
+  SELECTORS: {
+    TODO: '.c-todo',
+    TITLE: '.c-todo__title',
+    LIST: '.c-todo__list',
+    ITEM: '.c-todo__item',
+    ITEM_CONTENT: '.c-todo__item-content',
+    ITEM_TEXT: '.c-todo__item-text',
+    ITEM_ACTIONS: '.c-todo__item-actions',
+    CHECKBOX: '.c-todo__checkbox',
+    DELETE_BUTTON: '.c-todo__delete-btn',
+    FORM: '.c-todo__form',
+    INPUT: '.c-todo__input',
+    ADD_BUTTON: '.c-todo__add-btn',
+  },
+  CLASSES: {
+    BASE: 'c-todo',
+    ITEM: 'c-todo__item',
+    COMPLETED: 'c-todo__item--completed',
+    SMALL: 'c-todo--sm',
+    LARGE: 'c-todo--lg',
+  },
+};
