+++ src/lib/constants/components/toggle.ts
@@ -1,0 +1,11 @@
+/**
+ * Toggle-specific constants
+ */
+export const TOGGLE = {
+  SELECTORS: {
+    TOGGLE: '.c-toggle',
+  },
+  CLASSES: {
+    IS_ON: 'is-on',
+  },
+};
