+++ src/lib/constants/components/upload.ts
@@ -1,0 +1,33 @@
+/**
+ * Upload-specific constants
+ */
+export const UPLOAD = {
+  SELECTORS: {
+    UPLOAD: '.c-upload',
+    INNER: '.c-upload__inner',
+    ICON: '.c-upload__icon',
+    TITLE: '.c-upload__title',
+    TEXT: '.c-upload__text',
+    BUTTON: '.c-upload__btn',
+    HELPER_TEXT: '.c-upload__helper-text',
+    LOADER: '.c-upload__loader',
+    LOADER_STATUS: '.c-upload__loader-status',
+    LOADER_TITLE: '.c-upload__loader-title',
+    LOADER_PROGRESS: '.c-upload__loader-progress',
+    LOADER_PAR: '.c-upload__loader-par',
+    LOADER_TIME: '.c-upload__loader-time',
+    LOADER_CONTROL: '.c-upload__loader-control',
+    LOADER_BAR: '.c-upload__loader-bar',
+    LOADER_CLOSE: '.c-upload__loader-close',
+  },
+  CLASSES: {
+    DISABLED: 'c-upload--disabled',
+    ERROR: 'c-upload--error',
+    SUCCESS: 'c-upload--success',
+    LOADING: 'c-upload--loading',
+    DRAGGING: 'c-upload--dragging',
+  },
+  ATTRIBUTES: {
+    PERCENTAGE: '--upload-loader-percentage',
+  },
+};
