+++ src/lib/constants/components/video-player.ts
@@ -1,0 +1,56 @@
+/**
+ * VideoPlayer component constants
+ */
+export const VIDEO_PLAYER = {
+  SELECTORS: {
+    VIDEO_PLAYER: '.c-video-player',
+    VIDEO: '.c-video-player__video',
+    CONTROLS: '.c-video-player__controls',
+    PROGRESS: '.c-video-player__progress',
+    VOLUME: '.c-video-player__volume',
+    SETTINGS: '.c-video-player__settings',
+  },
+  CLASSES: {
+    BASE: 'c-video-player',
+    VIDEO: 'c-video-player__video',
+    YOUTUBE: 'c-video-player--youtube',
+    LOADING: 'c-video-player__loading',
+    SPINNER: 'c-video-player__spinner',
+    CONTROLS: 'c-video-player__controls',
+    CONTROLS_VISIBLE: 'c-video-player__controls--visible',
+    PROGRESS_CONTAINER: 'c-video-player__progress-container',
+    PROGRESS_BAR: 'c-video-player__progress-bar',
+    PROGRESS_BUFFERED: 'c-video-player__progress-buffered',
+    PROGRESS_PLAYED: 'c-video-player__progress-played',
+    PROGRESS_THUMB: 'c-video-player__progress-thumb',
+    CONTROLS_ROW: 'c-video-player__controls-row',
+    CONTROLS_LEFT: 'c-video-player__controls-left',
+    CONTROLS_RIGHT: 'c-video-player__controls-right',
+    CONTROL_BUTTON: 'c-video-player__control-button',
+    VOLUME_CONTAINER: 'c-video-player__volume-container',
+    VOLUME_SLIDER: 'c-video-player__volume-slider',
+    VOLUME_BAR: 'c-video-player__volume-bar',
+    VOLUME_FILL: 'c-video-player__volume-fill',
+    TIME_DISPLAY: 'c-video-player__time-display',
+    SETTINGS_CONTAINER: 'c-video-player__settings-container',
+    SETTINGS_MENU: 'c-video-player__settings-menu',
+    SETTINGS_TABS: 'c-video-player__settings-tabs',
+    SETTINGS_TAB: 'c-video-player__settings-tab',
+    SETTINGS_TAB_ACTIVE: 'c-video-player__settings-tab--active',
+    SETTINGS_CONTENT: 'c-video-player__settings-content',
+    SETTINGS_OPTIONS: 'c-video-player__settings-options',
+    SETTINGS_OPTION: 'c-video-player__settings-option',
+    SETTINGS_OPTION_ACTIVE: 'c-video-player__settings-option--active',
+    AMBIENT: 'c-video-player--ambient',
+    AMBIENT_CANVAS: 'c-video-player__ambient-canvas',
+    GLASS: 'c-video-player--glass',
+    GLASS_OVERLAY: 'c-video-player__glass-overlay',
+    GLASS_CONTENT: 'c-video-player__glass-content',
+  },
+  DEFAULTS: {
+    CONTROLS_TIMEOUT: 3000,
+    VOLUME: 1,
+    PLAYBACK_RATE: 1,
+    ASPECT_RATIO: '16:9',
+  },
+};
